--- Client/src/hooks/useLocalStorage.jsx
@@ -1,27 +1,0 @@
-import { useState } from "react";
-
-export const useLocalStorage = (key, initialValue) => {
-
-    // se utiliza como un useState pero en vez de guardar valores en el estado del componente
-    // lo hace en el localStorage
-
-  const [storedValue, setStoredValue] = useState(() => {
-    try {
-      const item = window.localStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      return initialValue;
-    }
-  });
-
-  const setValue = (value) => {
-    try {
-      setStoredValue(value);
-      window.localStorage.setItem(key, JSON.stringify(value));
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  return [storedValue, setValue]
-};
